feat: 🎸 Parse Elements Map
Finishing parsing elements and map for merging into main branch
