+++ .gitignore
@@ -1,6 +1,7 @@
 cub3d
 package.json
 maps/
+minilibx-linux/test/mlx-test
 *.swp
 *.out
 *.o
