+++ assets/pavingstone.xpm
