--- teste.cub
@@ -1,22 +1,0 @@
-NO assets/rusts.xpm
-SO assets/rusts.xpm
-WE assets/rusts.xpm
-EA assets/rusts.xpm
-F 220,100,0
-C 225,30,0
-
-
-        1111111111111111111111111
-        1000000000110000000000001
-        1011000001110000000000001
-        1001000000000000000000001
-111111111011000001110000000000001
-100000000011000001110111111111111
-11110111111111011100000010001
-11110111111111011101010010001
-11000001110101011100000010001
-1000001 100000001100000010001
-10000001000000001101010010001
-11000001110101011111011110N0111
-11110111 1110101 101111010001
-11111111 1111111 11 111111111
